vertex split not ok yet because there is a problem on the edge collapse : an edge and his twin has the same starting vertex. See and uncomment l. 157 of EdgeCollapse.cpp

+++ src/Core/Index/IndexMap.inl
@@ -284,7 +284,7 @@ inline bool IndexMap< T >::contain( const Index& idx ) const {
 
 template < typename T >
 inline bool IndexMap< T >::compact() const {
-    return ( m_free.front() > m_data.size() );
+    return ( size_t(m_free.front()) > m_data.size() );
 }
